refac: ã: log앱 내 리팩토링
