+++ src/main/java/com/likelion/fourthlinethon/team1/cooltime/log/repository/DailyLogRepository.java
@@ -15,17 +15,13 @@ import java.util.Optional;
 @Repository
 public interface DailyLogRepository extends JpaRepository<DailyLog, Long> {
 
-    // ✅ 특정 날짜의 기록 존재 여부
     Optional<DailyLog> findByUserAndDate(User user, LocalDate date);
     boolean existsByUserAndDate(User user, LocalDate date);
 
-    // ✅ 기간별 기록 조회
     List<DailyLog> findAllByUserAndDateBetween(User user, LocalDate startDate, LocalDate endDate);
 
-    // ✅ 특정 유저의 총 기록 수 반환
     long countByUser(User user);
 
-    // ✅ 통계용: 특정 기간의 미룸/성공 기록 수 집계
     @Query("""
             SELECT
                 count(dl) as total,
@@ -41,7 +37,6 @@ public interface DailyLogRepository extends JpaRepository<DailyLog, Long> {
             @Param("endDate") LocalDate endDate
     );
 
-    // ✅ [추가] 특정 연도와 월 기준으로 일별 로그 조회 (캘린더용)
     @Query("""
         select dl
         from DailyLog dl
