+++ src/main/java/com/likelion/fourthlinethon/team1/cooltime/user/repository/UserRepository.java
@@ -13,6 +13,5 @@ public interface UserRepository extends JpaRepository<User, Long> {
 
     boolean existsByNickname(String nickname);
 
-    // 리프레시 토큰으로 사용자 조회 (JWT 토큰 재발급 시 사용)
     Optional<User> findByRefreshToken(String refreshToken);
 }
